Move send_mass_html_mail function to registrations utils
